Update network schema tests
